split up install module
